More logic moved into services. TODO: Tests and documentation

+++ src/main/java/com/example/demo/Request/RequestService.java
@@ -1,13 +1,20 @@
 package com.example.demo.Request;
 
+import com.example.demo.Authentication.AuthenticationService;
 import com.example.demo.Project.Project;
-import com.example.demo.Volunteer.Volunteer;
+import com.example.demo.Project.ProjectDTO;
+import com.example.demo.Project.ProjectMapper;
 import com.example.demo.Project.ProjectRepository;
+import com.example.demo.Volunteer.Volunteer;
+import com.example.demo.Volunteer.VolunteerDTO;
+import com.example.demo.Volunteer.VolunteerMapper;
+import com.example.demo.Volunteer.VolunteerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * Service responsible for request operations
@@ -21,14 +28,39 @@ public class RequestService {
     @Autowired
     private ProjectRepository projectRepository;
 
+    @Autowired
+    private VolunteerService volunteerService;
+
+    @Autowired
+    private RequestMapper requestMapper;
+
+    @Autowired
+    private VolunteerMapper volunteerMapper;
+
+    @Autowired
+    private AuthenticationService authenticationService;
+
+    @Autowired
+    private ProjectMapper projectMapper;
+
+
     /**
      * Searches for requests
      *
      * @return List of found requests
      */
-    public List<VolunteerRequest> searchAllRequests() {
+    public Optional<List<RequestDTO>> searchAllRequests() {
 
-        return requestRepository.findAll();
+        List<VolunteerRequest> foundRequests = requestRepository.findAll();
+
+        if (!foundRequests.isEmpty()) {
+            List<RequestDTO> requestDTOs = foundRequests.stream()
+                    .map(request -> requestMapper.mapRequestToDTO(request))
+                    .collect(Collectors.toList());
+            return Optional.of(requestDTOs);
+        }
+
+        return Optional.empty();
     }
 
     /**
@@ -42,14 +74,42 @@ public class RequestService {
         return requestRepository.findById(id);
     }
 
+    public Optional<VolunteerDTO> getRequestSender(VolunteerRequest request) {
+
+        if (this.isVolunteerSenderOrReceiver(request, volunteerService.getLoggedVolunteer()) || authenticationService.checkIfAdmin(volunteerService.getLoggedVolunteer())) {
+
+            return Optional.of(volunteerMapper.mapVolunteerToDTO(request.getRequestSender()));
+        }
+
+        return Optional.empty();
+    }
+
+    public Optional<VolunteerDTO> getRequestReceiver(VolunteerRequest request) {
+
+        if (this.isVolunteerSenderOrReceiver(request, volunteerService.getLoggedVolunteer()) || authenticationService.checkIfAdmin(volunteerService.getLoggedVolunteer())) {
+
+            return Optional.of(volunteerMapper.mapVolunteerToDTO(request.getRequestSender()));
+        }
+
+        return Optional.empty();
+    }
+
     /**
      * Deletes requests
      *
      * @param request Deleted request
      */
-    public void deleteRequest(VolunteerRequest request) {
+    public Optional<RequestDTO> deleteRequest(VolunteerRequest request) {
 
-        requestRepository.delete(request);
+        if (authenticationService.checkIfAdmin(volunteerService.getLoggedVolunteer())) {
+
+            requestRepository.delete(request);
+
+            RequestDTO requestDTO = requestMapper.mapRequestToDTO(request);
+
+            return Optional.of(requestDTO);
+        }
+        return Optional.empty();
     }
 
     /**
@@ -102,18 +162,26 @@ public class RequestService {
     /**
      * Creates request
      *
-     * @param projectOwner Volunteer object assumed to receive request
-     * @param loggedUser   Volunteer object assumed to send request
      * @param foundProject Project that is regarded in request
      * @return Created request
      */
-    public VolunteerRequest createRequest(Volunteer projectOwner, Volunteer loggedUser, Project foundProject) {
+    public RequestDTO createRequest(Project foundProject) {
 
-        VolunteerRequest newRequest = VolunteerRequest.builder().requestReceiver(projectOwner).requestSender(loggedUser).requestedProject(foundProject).status(RequestStatus.PENDING).build();
+        VolunteerRequest newRequest = VolunteerRequest.builder()
+                .requestReceiver(foundProject.getOwnerVolunteer())
+                .requestSender(this.volunteerService.getLoggedVolunteer())
+                .requestedProject(foundProject)
+                .status(RequestStatus.PENDING)
+                .build();
 
         requestRepository.save(newRequest);
 
-        return newRequest;
+        return requestMapper.mapRequestToDTO(newRequest);
+    }
+
+    public Optional<ProjectDTO> getRequestedProject(VolunteerRequest request) {
+
+        return Optional.of(projectMapper.mapProjectToDTO(request.getRequestedProject()));
     }
 
     /**
@@ -122,14 +190,16 @@ public class RequestService {
      * @param request Request that will be declined
      * @return Optional value that contains result of operation
      */
-    public Optional<VolunteerRequest> declineRequest(VolunteerRequest request) {
+    public Optional<RequestDTO> declineRequest(VolunteerRequest request) {
 
-        if (this.hasPendingStatus(request)) {
+        if ((this.isVolunteerReceiver(request, volunteerService.getLoggedVolunteer()) || authenticationService.checkIfAdmin(volunteerService.getLoggedVolunteer())) && volunteerService.getLoggedVolunteer().getOwnedProjects().contains(request.getRequestedProject()) && this.hasPendingStatus(request)) {
 
             request.setStatus(RequestStatus.DECLINED);
             requestRepository.save(request);
 
-            return Optional.of(request);
+            RequestDTO requestDTO = requestMapper.mapRequestToDTO(request);
+
+            return Optional.of(requestDTO);
         } else {
 
             return Optional.empty();
@@ -140,21 +210,21 @@ public class RequestService {
      * Accepts request
      *
      * @param request   Request that will be accepted
-     * @param volunteer Volunteer that will be added to project
-     * @param project   Project that will be modified
      * @return Optional value that contains result of operation
      */
-    public Optional<VolunteerRequest> acceptRequest(VolunteerRequest request, Volunteer volunteer, Project project) {
+    public Optional<RequestDTO> acceptRequest(VolunteerRequest request) {
 
-        if (this.hasPendingStatus(request)) {
+        if ((this.isVolunteerReceiver(request, volunteerService.getLoggedVolunteer()) || authenticationService.checkIfAdmin(volunteerService.getLoggedVolunteer())) && volunteerService.getLoggedVolunteer().getOwnedProjects().contains(request.getRequestedProject()) && this.hasPendingStatus(request)) {
 
-            project.addVolunteerToProject(volunteer);
+            request.getRequestedProject().addVolunteerToProject(request.getRequestSender());
             request.setStatus(RequestStatus.ACCEPTED);
 
-            projectRepository.save(project);
+            projectRepository.save(request.getRequestedProject());
             requestRepository.save(request);
 
-            return Optional.of(request);
+            RequestDTO requestDTO = requestMapper.mapRequestToDTO(request);
+
+            return Optional.of(requestDTO);
         } else {
 
             return Optional.empty();
